+++ src/net/master/ControlClient.cpp
@@ -354,7 +354,7 @@ int ControlClient::SendCtrlData()
 	int src = 0;
 
 	//为每个寄存器设置单独地址
-	zlog_info(Util::m_zlog,"下发数据大小%d",data.size());
+	zlog_info(Util::m_zlog,"下发数据大小%d", (int)data.size());
 	for (i = 0; i < data.size(); i++)
 	{
 		realData = (RealData*) data[i];
